+++ CHANGELOG.md
@@ -4,6 +4,7 @@ It is based on Tidepool Blip 1.27.
 
 ## 3.5.0-beta - 2024-09-03
 ### Added
+- YLP-2408 Display basal safety profile in Devices view
 - YLP-3128 As an HCP, I want to see which device is used by each patient
 
 ## 3.4.1 - 2024-09-25